Time: 1 ms (99.93%), Space: 77.2 MB (49.04%) - LeetHub

+++ 0053-maximum-subarray/0053-maximum-subarray.java
@@ -1,15 +1,13 @@
 class Solution {
     public int maxSubArray(int[] nums) {
-        long max=Long.MIN_VALUE;
-
-        long sum=0;
-
-        for(int i=0; i<nums.length; i++){
-            sum+=nums[i];
-
-            if(sum>max) max=sum;
-
-            if(sum<0) sum=0;
+        long max = Integer.MIN_VALUE;
+        int sum = 0;
+        for (int i = 0; i < nums.length; i++) {
+            sum += nums[i];
+            if (sum > max)
+                max = sum;
+            if (sum < 0)
+                sum = 0;
         }
         return (int) max;
     }
